refactor(challenge): 패키지 구조 재정비
