--- src/main/java/com/sopt/cherrish/domain/challenge/application/service/HomecareRoutineService.java
@@ -1,28 +1,0 @@
-package com.sopt.cherrish.domain.challenge.application.service;
-
-import java.util.Arrays;
-import java.util.List;
-
-import org.springframework.stereotype.Service;
-
-import com.sopt.cherrish.domain.challenge.application.dto.response.HomecareRoutineResponseDto;
-import com.sopt.cherrish.domain.challenge.domain.model.HomecareRoutine;
-
-/**
- * 홈케어 루틴 조회 서비스
- */
-@Service
-public class HomecareRoutineService {
-
-	/**
-	 * 모든 홈케어 루틴 목록 조회
-	 *
-	 * @return 홈케어 루틴 목록
-	 */
-	public List<HomecareRoutineResponseDto> getAllHomecareRoutines() {
-
-		return Arrays.stream(HomecareRoutine.values())
-			.map(HomecareRoutineResponseDto::from)
-			.toList();
-	}
-}
