+++ src/main/java/com/sopt/cherrish/domain/challenge/core/exception/ChallengeException.java
@@ -1,0 +1,10 @@
+package com.sopt.cherrish.domain.challenge.core.exception;
+
+import com.sopt.cherrish.global.exception.BaseException;
+
+public class ChallengeException extends BaseException {
+
+	public ChallengeException(ChallengeErrorCode errorCode) {
+		super(errorCode);
+	}
+}
