--- src/main/java/com/sopt/cherrish/domain/challenge/exception/ChallengeException.java
@@ -1,10 +1,0 @@
-package com.sopt.cherrish.domain.challenge.exception;
-
-import com.sopt.cherrish.global.exception.BaseException;
-
-public class ChallengeException extends BaseException {
-
-	public ChallengeException(ChallengeErrorCode errorCode) {
-		super(errorCode);
-	}
-}
